fix(sales.types): fixed query Sale with price to Float

+++ src/types/sales.types.js
@@ -9,11 +9,11 @@ module.exports = `#graphql
         product: String
         suplier: String
         quantity: Int
-        price: Int
-        total: Int
-        costPrice: Int
+        price: Float
+        total: Float
+        costPrice: Float
         description: String
-        profit: Int
+        profit: Float
         percentage: Float
         date: String
     }
